# sdk/src/qorstack-swagger.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Qorstack Report API",
    "description": "API for generating PDF reports from Word templates",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:5001/"
    }
  ],
  "paths": {
...
    "/analytics/dashboard-summary": {
      "get": {
        "tags": ["Analytics"],
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "7D"
            }
          },
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "schema": {
              "format": "date-time"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "schema": {
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryDto"
                }
              }
            }
          }
        },
        "security": [{}]
      }
    },
    "/analytics/contributions": {
      "get": {
        "tags": ["Analytics"],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": ["integer", "string"],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributionDataDto"
                }
              }
            }
          }
        },
        "security": [{}]
      }
    },
    "/analytics/usage": {
      "get": {
        "tags": ["Analytics"],
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "7D"
            }
          },
          {
            "name": "groupBy",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "day"
            }
          },
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "schema": {
              "format": "date-time"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "schema": {
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageDataDto"
                }
              }
            }
          }
        },
        "security": [{}]
      }
    },
    "/analytics/usage/hourly": {
      "get": {
        "tags": ["Analytics"],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HourlyUsageDto"
                  }
                }
              }
            }
          }
        },
        "security": [{}]
      }
    },
    "/analytics/usage/weekly": {
      "get": {
        "tags": ["Analytics"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeeklyUsageDto"
                  }
                }
              }
            }
          }
        },
        "security": [{}]
      }
    },
    "/analytics/templates": {
      "get": {
        "tags": ["Analytics"],
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "schema": {
              "format": "date-time"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "schema": {
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplatePerformanceDto"
                  }
                }
              }
            }
          }
        },
        "security": [{}]
      }
    },
    "/analytics/generations": {
      "get": {
        "tags": ["Analytics"],
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "templateKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "schema": {
              "format": "date-time"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "schema": {
              "format": "date-time"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": ["integer", "string"],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": ["integer", "string"],
              "format": "int32",
              "default": 15
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "createdDatetime"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedListOfGenerationDto"
                }
              }
            }
          }
        },
        "security": [{}]
      }
    },
...
    "/templates": {
      "post": {
        "tags": ["Templates"],
        "summary": "Upload a new template",
        "description": "Upload a template file (DOCX or XLSX). Name is required via query parameter, file is required via multipart form data.",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "templateKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "isAutoGeneratedVariable",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [{}]
      },
      "get": {
        "tags": ["Templates"],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": ["integer", "string"],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": ["integer", "string"],
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedListOfTemplateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [{}]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BusiestDayDto": {
        "required": ["date", "count"],
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          }
        }
      },
...
      "ContributionDataDto": {
        "required": [
          "totalContributions",
          "growthPercent",
          "busiestDay",
          "activeDays",
          "dailyData"
        ],
        "type": "object",
        "properties": {
          "totalContributions": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "growthPercent": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": ["number", "string"],
            "format": "double"
          },
          "busiestDay": {
            "$ref": "#/components/schemas/BusiestDayDto"
          },
          "activeDays": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "dailyData": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyContributionDto"
            }
          }
        }
      },
...
      "DailyContributionDto": {
        "required": ["date", "count"],
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          }
        }
      },
      "DashboardSummaryDto": {
        "required": [
          "totalGenerated",
          "totalGeneratedTrend",
          "activeProjects",
          "maxProjects",
          "successRate",
          "successRateTrend",
          "totalTemplates",
          "templateBreakdown"
        ],
        "type": "object",
        "properties": {
          "totalGenerated": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "totalGeneratedTrend": {
            "type": "array",
            "items": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": ["integer", "string"],
              "format": "int32"
            }
          },
          "activeProjects": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "maxProjects": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "successRate": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": ["number", "string"],
            "format": "double"
          },
          "successRateTrend": {
            "type": "array",
            "items": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": ["number", "string"],
              "format": "double"
            }
          },
          "totalTemplates": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "templateBreakdown": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateBreakdownDto"
            }
          }
        }
      },
...
      "GenerationDto": {
        "required": [
          "id",
          "templateName",
          "templateKey",
          "type",
          "status",
          "createdDatetime",
          "durationMs",
          "fileSizeBytes",
          "errorMessage",
          "downloadUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "templateName": {
            "type": ["null", "string"]
          },
          "templateKey": {
            "type": ["null", "string"]
          },
          "type": {
            "type": ["null", "string"]
          },
          "status": {
            "type": "string"
          },
          "createdDatetime": {
            "format": "date-time"
          },
          "durationMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["null", "integer", "string"],
            "format": "int64"
          },
          "fileSizeBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["null", "integer", "string"],
            "format": "int64"
          },
          "errorMessage": {
            "type": ["null", "string"]
          },
          "downloadUrl": {
            "type": ["null", "string"]
          }
        }
      },
...
      "HourlyUsageDto": {
        "required": ["hour", "count"],
        "type": "object",
        "properties": {
          "hour": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          }
        }
      },
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "PaginatedListOfGenerationDto": {
        "required": ["items", "pageNumber"],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenerationDto"
            }
          },
          "pageNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        }
      },
...
      "PaginatedListOfTemplateResponse": {
        "required": ["items", "pageNumber"],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateResponse"
            }
          },
          "pageNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": ["null", "string"]
          },
          "title": {
            "type": ["null", "string"]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["null", "integer", "string"],
            "format": "int32"
          },
          "detail": {
            "type": ["null", "string"]
          },
          "instance": {
            "type": ["null", "string"]
          }
        }
      },
...
      "TemplateBreakdownDto": {
        "required": ["type", "count"],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          }
        }
      },
      "TemplateDetailResponse": {
        "required": ["id", "userId", "templateKey", "name"],
        "type": "object",
        "properties": {
          "fileSandboxLastTestPresigned": {
            "type": ["null", "string"]
          },
          "activeVersion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TemplateVersionResponse"
              }
            ]
          },
          "allVersions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateVersionResponse"
            }
          },
          "sandboxPayload": {
            "type": ["null", "string"]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": ["null", "string"],
            "format": "uuid"
          },
          "templateKey": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdBy": {
            "type": ["null", "string"]
          },
          "createdDatetime": {
            "format": "date-time"
          },
          "updatedBy": {
            "type": ["null", "string"]
          },
          "updatedDatetime": {
            "format": "date-time"
          }
        }
      },
      "TemplatePerformanceDto": {
        "required": [
          "templateKey",
          "templateName",
          "projectName",
          "type",
          "totalGenerations",
          "avgDurationMs",
          "avgFileSizeBytes",
          "errorCount",
          "successRate",
          "errorRate",
          "lastGeneratedAt"
        ],
        "type": "object",
        "properties": {
          "templateKey": {
            "type": "string"
          },
          "templateName": {
            "type": "string"
          },
          "projectName": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "totalGenerations": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "avgDurationMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": ["number", "string"],
            "format": "double"
          },
          "avgFileSizeBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int64"
          },
          "errorCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "successRate": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": ["number", "string"],
            "format": "double"
          },
          "errorRate": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": ["number", "string"],
            "format": "double"
          },
          "lastGeneratedAt": {
            "format": "date-time"
          }
        }
      },
      "TemplateResponse": {
        "required": ["id", "userId", "templateKey", "name"],
        "type": "object",
        "properties": {
          "activeVersion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TemplateVersionResponse"
              }
            ]
          },
          "allVersions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateVersionResponse"
            }
          },
          "sandboxPayload": {
            "type": ["null", "string"]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": ["null", "string"],
            "format": "uuid"
          },
          "templateKey": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdBy": {
            "type": ["null", "string"]
          },
          "createdDatetime": {
            "format": "date-time"
          },
          "updatedBy": {
            "type": ["null", "string"]
          },
          "updatedDatetime": {
            "format": "date-time"
          }
        }
      },
      "TemplateVersionResponse": {
        "required": ["id", "templateId", "version", "filePath"],
        "type": "object",
        "properties": {
          "previewFilePathPresigned": {
            "type": ["null", "string"]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "templateId": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "filePath": {
            "type": "string"
          },
          "status": {
            "type": ["null", "string"]
          },
          "previewFilePath": {
            "type": ["null", "string"]
          },
          "sandboxPayload": {
            "type": ["null", "string"]
          },
          "sandboxFilePath": {
            "type": ["null", "string"]
          },
          "createdBy": {
            "type": ["null", "string"]
          },
          "createdDatetime": {
            "format": "date-time"
          },
          "updatedBy": {
            "type": ["null", "string"]
          },
          "updatedDatetime": {
            "format": "date-time"
          }
        }
      },
...
      "UsageDataDto": {
        "required": ["range", "groupBy", "totalVolume", "data"],
        "type": "object",
        "properties": {
          "range": {
            "type": "string"
          },
          "groupBy": {
            "type": "string"
          },
          "totalVolume": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UsageDataPointDto"
            }
          }
        }
      },
      "UsageDataPointDto": {
        "required": ["date", "count", "breakdown"],
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          },
          "breakdown": {
            "type": "object",
            "additionalProperties": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": ["integer", "string"],
              "format": "int32"
            }
          }
        }
      },
...
      "WeeklyUsageDto": {
        "required": ["date", "count"],
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["integer", "string"],
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "API Key authentication. Enter your API key in the format: rdx_xxxxx",
        "name": "X-API-Key",
        "in": "header"
      }
    }
  },
...
}